+++ HelloASL/composeApp/src/commonMain/kotlin/ca/uwaterloo/helloasl/domain/learningModel/Module.kt
@@ -1,7 +1,7 @@
 package ca.uwaterloo.helloasl.domain.learningModel
 
 data class Module(
-    val moduleId: Int,
+    val moduleId: Long,
     val title: String,
     val category: String? = null
 )
